move widgets to layout to reduce bundle size
